not sure stuff works kinda

+++ Project/src/blueBookTab/BlueBookPreview.java
@@ -2,6 +2,7 @@ package blueBookTab;
 
 import java.awt.Color;
 import java.awt.Container;
+import java.awt.Desktop;
 import java.awt.Dimension;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -11,7 +12,6 @@ import java.nio.file.Path;
 import javax.swing.JButton;
 import javax.swing.JDialog;
 import javax.swing.JFrame;
-import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
@@ -19,6 +19,7 @@ import javax.swing.JToolBar;
 import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingConstants;
 import net.miginfocom.swing.MigLayout;
+import program.ResourceManager;
 import utilities.pdf.PDFView;
 import utilities.ui.SwingHelper;
 //-----------------------------------------------------------------------------
@@ -28,7 +29,10 @@ import utilities.ui.SwingHelper;
  * can be edited, saved, printed, and emailed.
  */
 public class BlueBookPreview extends JDialog implements ActionListener {
-private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
+	private static String DELETE_ACTION = "delete";
+	private static String PRINT_ACTION = "print";
+	private static String EMAIL_ACTION = "email";
 	private String name, dob, affili, address, crimeDescrip, narrative;
 	private String date, time, location, caseNum, status, weapon;
 	private Integer bbID=null;
@@ -44,6 +48,8 @@ private static final long serialVersionUID = 1L;
 	/** JDOC */
 	JPanel dialogPanel;
 	PDFView pdfv;
+	ResourceManager rm;
+	    private Desktop.Action action = Desktop.Action.OPEN;
 //-----------------------------------------------------------------------------
 	/**
 	 * Generates the <code>BlueBookPreview</code> window  with all necessary fields 
@@ -52,13 +58,13 @@ private static final long serialVersionUID = 1L;
 	 * @param parent - JDOC
 	 * @param bbEntry - JDOC
 	 */
-	public BlueBookPreview(JFrame parent, BlueBookTab bbTab, BlueBookEntry bbEntry){
-			super(parent, "BlueBook Entry", true);
+	public BlueBookPreview(ResourceManager rm, BlueBookTab bbTab, BlueBookEntry bbEntry){
+			super(rm.getGuiParent(), "BlueBook Entry", true);
 //DEBUG
 System.out.println("BlueBookPreview: constructor ");
 			//BlueBookEntry object to load info from
 			this.bbEntry = bbEntry;
-			this.parent = parent;
+			this.rm=rm;
 			this.bbTab=bbTab;
 			
 			//Set the size of the page
@@ -103,9 +109,12 @@ System.out.println("BlueBookPreview: constructor ");
 		    
 		    
 			//Create a PDF from the given BlueBookEntry based on the form template
-			String filename = bbEntry.saveToFileSystem();
+			String filename="";
+			try{
+				filename = bbEntry.createPdf(bbEntry.createNewUniqueFilename());
+			} catch(Exception e){ e.printStackTrace(); }
 			
-		    JScrollPane scroller = new JScrollPane();
+			JScrollPane scroller = new JScrollPane();
 		    
 			pdfv = new PDFView(filename, scroller, createButtonsPanel());
 			Container contentPane = this.getContentPane();
@@ -116,13 +125,11 @@ System.out.println("BlueBookPreview: constructor ");
 	/**
 	 * JDOC
 	 */
-	//private JToolBar createToolbar(){
 		private JPanel createButtonsPanel(){
 		JToolBar toolbar = new JToolBar(SwingConstants.HORIZONTAL);
 		toolbar.setFloatable(false);
 		
 		JPanel panel = new JPanel();
-		
 		
 		//Cancel button
 		JButton cancelButton = SwingHelper.createImageButton("Cancel", "icons/cancel_32.png");
@@ -146,6 +153,7 @@ System.out.println("BlueBookPreview: constructor ");
 	    JButton deleteButton = SwingHelper.createImageButton("Delete", 
 	    		"icons/delete_32.png");
 	    deleteButton.setToolTipText("Delete BlueBookEntry");
+	    deleteButton.setActionCommand(DELETE_ACTION);
 	    deleteButton.addActionListener(this);
 	    
 	    //Edit button
@@ -165,21 +173,15 @@ System.out.println("BlueBookPreview: constructor ");
 	    JButton printButton = 
 	    		SwingHelper.createImageButton("Print", "icons/print_32.png");
 	    printButton.setToolTipText("Print this BlueBookEntry document");
-	    printButton.addActionListener(new ActionListener( ) {
-	    	public void actionPerformed(ActionEvent e) {
-	 //   		PrintHelper ph = new PrintHelper(dialogPanel);
-	    	}
-	    });
+	    printButton.setActionCommand(PRINT_ACTION);
+	    printButton.addActionListener(this);
+	    
 	    
 	    //Email button
 	    JButton emailButton = SwingHelper.createImageButton("Email", "icons/email32.png");
 	    emailButton.setToolTipText("Email this BlueBookEntry document");
-	    emailButton.addActionListener(new ActionListener( ) {
-	    	public void actionPerformed(ActionEvent e) {
-	    		//TODO implement email
-	    	}
-	    });
-	    
+	    emailButton.setActionCommand(EMAIL_ACTION);
+	    emailButton.addActionListener(this);
 	    /*
 	    JPanel saveAndCancelButtonsPanel = new JPanel();
 	    saveAndCancelButtonsPanel.add(saveButton, "tag ok, dock west");
@@ -232,8 +234,16 @@ System.out.println("BlueBookPreview: constructor ");
 	}
 //-----------------------------------------------------------------------------
 	public void actionPerformed(ActionEvent ev) {
-		//attempt to delete the currently displayed BlueBookEntry & close this dialog
-		deleteEntryAndClose();	
+		if(ev.getActionCommand().equals(DELETE_ACTION)){
+			//attempt to delete the currently displayed BlueBookEntry & close this dialog
+			deleteEntryAndClose();	
+		} else if(ev.getActionCommand().equals(PRINT_ACTION)){
+			rm.launchDefaultApplication(bbEntry.filename);
+		} else if(ev.getActionCommand().equals(EMAIL_ACTION)){
+			Desktop.Action action = Desktop.Action.OPEN;
+			//rm.launchMail(ev);
+			
+		}
 	}
 //-----------------------------------------------------------------------------
 	/**
@@ -266,6 +276,7 @@ System.out.println("BlueBookPreview: constructor ");
 		
 	}
 //-----------------------------------------------------------------------------
+	
 }
